CDDSO-449 Minor changes to mip_convert
* Organise imports
* Split test_request.py into requested_variables.py and test_mip_table.py
* Move unused code
