+++ cdds/cdds/prepare/common.py
@@ -9,7 +9,8 @@ import configparser
 
 from collections import defaultdict
 from cdds.common.mappings.ancils import remove_ancils_from_mapping
-from mip_convert.requested_variables import get_model_to_mip_mappings, get_variable_model_to_mip_mapping
+from mip_convert.requested_variables import get_variable_model_to_mip_mapping
+from mip_convert.mip_table import get_model_to_mip_mappings
 
 
 def retrieve_mappings(data_request_variables, mip_era, model_id):
